{{cookiecutter.project_name}}/requirements.txt
beautifulsoup4==4.4.1
boto==2.41.0
botocore==1.4.37
click==6.6
Collectfast==0.2.3
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.9.7
django-anymail==0.4.2
django-modelcluster==2.0
django-storages==1.4.1
django-taggit==0.18.3
django-toolbelt==0.0.1
django-treebeard==4.0.1
djangorestframework==3.4.0
docutils==0.12
gunicorn==19.6.0
html5lib==0.9999999
jmespath==0.9.0
ordereddict==1.1
Pillow==3.3.0
psycopg2==2.6.2
python-dateutil==2.5.3
python-dotenv==0.5.1
pytz==2016.6.1
requests==2.10.0
six==1.10.0
static3==0.7.0
Unidecode==0.4.19
wagtail==1.5.2
webencodings==0.5
Willow==0.3.1
